# api/openapi.yaml
openapi: "3.1.0"
info:
  title: openByte Speed Test API
  version: "1.0"
  description: |
    REST API for the openByte network speed test server.
    All API endpoints are under `/api/v1/` except `/health`.
    All error responses return `{"error": "message"}`.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/download:
    get:
      summary: Download speed test (streams binary data)
      operationId: download
      tags: [SpeedTest]
      parameters:
        - name: duration
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 300
            default: 10
          description: Duration in seconds
        - name: chunk
          in: query
          schema:
            type: integer
            minimum: 65536
            maximum: 4194304
            default: 1048576
          description: Chunk size in bytes
      responses:
        "200":
          description: Binary data stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServerBusy"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerBusy:
      description: Server at capacity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...